+++ examples/do_signin.py
@@ -8,7 +8,7 @@ parent_path = os.path.dirname(
     )
 )
 if parent_path not in sys.path:
-    sys.path.append(parent_path)
+    sys.path.insert(0, parent_path)
 
 # isort: split
 from pyonepassword import OP  # noqa: E402
@@ -16,22 +16,19 @@ from pyonepassword.api.authentication import (  # noqa: E402
     EXISTING_AUTH_AVAIL,
     EXISTING_AUTH_REQD
 )
-from pyonepassword.api.exceptions import OPNotSignedInException  # noqa: E402
+from pyonepassword.api.exceptions import \
+    OPAuthenticationException  # noqa: E402
 
 
-def do_signin(vault=None, op_path="op", use_existing_session=False, account=None):
-    existing_auth = EXISTING_AUTH_AVAIL
-    if use_existing_session:
-        existing_auth = EXISTING_AUTH_REQD
-
+def do_signin(vault=None, op_path="op", existing_auth=EXISTING_AUTH_AVAIL, account=None, logger=None):
     # Let's check If biometric is enabled
     # If so, no need to provide a password
     uses_biometric = OP.uses_biometric(op_path=op_path)
     try:
         op = OP(vault=vault, op_path=op_path,
-                existing_auth=existing_auth, password_prompt=False, account=account)
-    except OPNotSignedInException as e:
-        if uses_biometric:
+                existing_auth=existing_auth, password_prompt=False, account=account, logger=logger)
+    except OPAuthenticationException as e:
+        if uses_biometric or existing_auth == EXISTING_AUTH_REQD:
             raise e
         # If you've already signed in at least once, you don't need to provide all
         # account details on future sign-ins. Just master password