+++ tests/fixtures/paths.py
@@ -18,6 +18,17 @@ ITEM_DELETE_MULTIPLE_TITLE_GLOB_STATE_CONFIG_PATH = Path(
 
 UNAUTH_RESP_DIRECTORY_PATH = Path(
     MOCK_OP_CONFIG_PATH, "unauth-response-directory.json")
+
+SVC_ACCT_RESP_DIRECTORY_PATH = Path(
+    MOCK_OP_CONFIG_PATH, "svc-acct-response-directory.json"
+)
+SVC_ACCT_CORRUPT_RESP_DIRECTORY_PATH = Path(
+    MOCK_OP_CONFIG_PATH, "svc-acct-corrupt-response-directory.json"
+)
+SVC_ACCT_REVOKED_RESP_DIRECTORY_PATH = Path(
+    MOCK_OP_CONFIG_PATH, "svc-acct-revoked-token-directory.json"
+)
+
 VALID_DATA_REGISTRY_PATH = Path(
     TEST_INPUT_DATA_PATH, "valid-data-registry.json")
 VALID_DATA_PATH = Path(TEST_INPUT_DATA_PATH, "valid-data")